Use int instead of long for seeds to be within JavaScript safe integer size

+++ backend/src/main/java/com/stefanlapointe/mathgym/model/ProblemResponse.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @Getter
 public class ProblemResponse {
     private String problemType;
-    private long seed;
+    private int seed;
     private String response;
 }
