+++ backend/src/main/java/com/stefanlapointe/mathgym/model/ProblemStatement.java
@@ -6,6 +6,6 @@ import lombok.Getter;
 @AllArgsConstructor
 @Getter
 public class ProblemStatement {
-    private long seed;
+    private int seed;
     private String statement;
 }
